enhancement: simplify user controller
