--- backend/app/controllers/auth/logout_controller.ts
@@ -1,10 +1,0 @@
-import type { HttpContext } from "@adonisjs/core/http";
-
-export default class LogoutController {
-  /**
-   * Handle user signout
-   */
-  async store({ request, response, auth }: HttpContext) {
-    await auth.use("web").logout();
-  }
-}
